requirements.txt
numpy
torch==1.4.0
torchvision==0.5
pandas
matplotlib
pytorch-lightning
albumentations
Pillow
